# ♻️Refactor: Simplify content negotiation code