@sixcolors 2024-02-19
I haven't had the chance to do a full review as I’m 🤧 and examining the PR on mobile. I'll conduct a more in-depth analysis before approving the PR. 

Based on what I’m seeing, the changes appear positive. The one thing I’m seeing is, considering that real-world scenarios often involve fewer than three parameters, the benchmarks suggest a more noticeable performance reduction. The performance trade-off seems to range from approximately 8% to 54% for these scenarios.