@nickajacks1 2024-02-19
Thanks for reviewing.
Indeed, there's an increased latency on my machine by a few dozen nanoseconds. To put that into perspective, it doesn't register as a blip on my laptop running a server that is doing nothing but content negotiation.  On more powerful machines running services that do actual work, I think we'd have a more difficult time finding a difference.  If nanoseconds mattered, you'd probably want to avoid content negotiation in the first place.  Individual benchmarks can be an excellent metric, but they are by definition very narrow in scope, and I find they can distract from more juicy areas if you're not careful.

That said, the biggest reason I did this refactor was for code clarity, so it's probably only worth the slice => map refactor if it leads to easier to understand code.  The switch to fasthttp.VisitHeaderParams can be done separately.

I can also repeat the same k6 benchmarks on some weaker armv7 embedded linux devices I have access to in order to see more exaggerated circumstances if we'd like to see more varied statistics.